fix: config renaming conflict
